Handle macro definitions moved to a different location
The USR generated for macro definitions includes location information
(file and line number) so if the USR is used as a key for these
definitions any change to their location would result in a removal and
addition being reported. Assume that API headers are sane as Clang does
for system headers and define a custom key for macro definitions.