Report relocation of declarations between headers
This is reported as two differences - one for the old location and
another for the new location.

Reporting relocations is useful in cases where consumers need to update
the header they are importing, and reporting the relocation for both the
old and new location makes reports easier to follow in cases where the
relocated symbol has also been modified.