Work around Clang issue with forward declaration a class or protocol
Clang returns the first declaration of an Objective-C class or protocol
as the canonical declaration. As a result a forward declaration of one
of these types that appears before the actual declaration is incorrectly
considered the canonical declaration. This breaks our isEqual: test
meant to ensure that we only consider the canonical declaration part of
the API. To work around the issue, check for one of these types by
comparing the cursor kind. Only the real canonical declaration will have
a cursor kind of one of these types.