Add --sdk option, add sysroot and deployment target automatically
As a convenience, support specifying the sysroot via --sdk. If a sysroot
is not included in the compiler arguments and the SDKROOT environment
variable is not set, use xcrun to locate the path to this SDK and add an
-isysroot argument.

Likewise, if an iOS SDK is specified add a deployment target if one has
not been specified via compiler arguments or environment variables.

This allows for simple use in the common case, but also use of explicit
sysroots and deployment targets if desired.