Use absolute paths when comparing declaration header paths
Through the command line arguments relative paths can be specified for
both the API paths and compiler search paths. When Clang includes a
header via a relative search path the source locations for all cursors
in that header are also relative to the current working directory.

To determine if a declaration has been relocated a comparison should be
made relative to the root of the API path, which may not be the current
working directory. Expand all paths to absolute paths in order to
consistently determine their common base.