Add support for comparing platform SDKs
All frameworks in the SDK are compared, with the exception of a few
macOS frameworks that cannot be parsed simply by including all headers
or the framework's umbrella header. The contents of /usr/include are
also not included, as these headers must be included in specific orders.
Using module maps would likely enable these to be included in the
future.

System frameworks are symlinked to a temporary directory and parsed with
a framework search path for this directory. This enables the fast
isInSystemHeader check to be used to exclude other system headers
outside the framework.

When a platform SDK or a framework within a platform SDK is given as one
of the API paths to compare appropriate sysroot and deployment targets
are added to the compiler arguments automatically if not explicitly
specified.