Exclude categories and class extension for classes in the same module
Children of the category or class extension are included, but the
category or extension itself is not. Instead, the extensions to the
class made via these declarations are reported as modifications against
the class. This better represents the effect that adding or removing a
category has on the API (e.g., extending the list of protocols a class
conforms to).

Categories on classes from other modules are still reported.