Add workaround for exclusion of builtin function declarations
Clang has an issue with declarations for functions that exist in its
builtin function database (e.g., NSLog, objc_msgSend). The canonical
cursor for these functions has an identical source location as the
declaration in the file we're parsing, but an extent that covers only
the function's name, so libclang does not consider them equal. Add a
workaround for this case and test to verify that such function
declarations are included.